fix: 2단계에 맞지않던 구간 제거
