--- src/main/java/nextstep/subway/domain/Section.java
@@ -1,55 +1,0 @@
-package nextstep.subway.domain;
-
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.ForeignKey;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-
-@Entity
-public class Section extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private int distance;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "up_station_id", foreignKey = @ForeignKey(name = "fk_section_to_upstation"))
-    private Station upStation;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "down_station_id", foreignKey = @ForeignKey(name = "fk_section_to_downstation"))
-    private Station downStation;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "line_id", foreignKey = @ForeignKey(name = "fk_section_to_line"))
-    private Line line;
-
-    public Section() {
-    }
-
-    public Section(Station upStation, Station downStation, int distance) {
-        this.upStation = upStation;
-        this.downStation = downStation;
-        this.distance = distance;
-    }
-
-    public static Section of(Station upStation, Station downStation, int distance) {
-        return new Section(upStation, downStation, distance);
-    }
-
-    public void addLine(Line line) {
-        this.line = line;
-    }
-
-    public Station getUpStation() {
-        return upStation;
-    }
-
-    public Station getDownStation() {
-        return downStation;
-    }
-}
